src/shared/pyproject.toml
[project]
name = "citizen-services-shared"
version = "0.1.0"
description = "Shared libraries for Citizen Services Portal"
requires-python = ">=3.11"
dependencies = [
    "azure-cosmos>=4.5.1",
    "azure-identity>=1.15.0",
    "pydantic>=2.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
